+++ source/Public/Invoke-BrevoCall.ps1
@@ -109,9 +109,10 @@ function Invoke-BrevoCall {
         }
     }
     catch {
-        $e = Get-Error -Newest 1
-        if ($e.TargetObject.Message) {
-            $e.TargetObject.Message | ConvertFrom-Json | Out-String | Write-Error
-        }
+        throw $_.Exception.Message
+        # $e = Get-Error -Newest 1
+        # if ($e.TargetObject.Message) {
+        #     $e.TargetObject.Message | ConvertFrom-Json | Out-String | Write-Error
+        # }
     }
 }